move abstract class (WP-854)
